update(types): make interface types exported
