+++ src/types.d.ts
@@ -1,0 +1,15 @@
+interface GithubProfile {
+  available: boolean
+  name?: string
+  avatar_url?: string
+  followers?: number
+  following?: number
+}
+export interface Quote {
+  id: string,
+  gradient_id: number,
+  username: string
+  github?: GithubProfile
+  text: string
+  author: string
+}
